+++ src/Repository/Model/Userflag.php
@@ -116,16 +116,22 @@ class Userflag extends database_object
     public static function garbage_collection($object_type = null, $object_id = null)
     {
         $types = array(
-            'song',
             'album',
             'album_disk',
             'artist',
-            'video',
-            'tvshow',
-            'tvshow_season',
+            'catalog',
+            'genre',
+            'label',
+            'live_stream',
             'playlist',
             'podcast',
-            'podcast_episode'
+            'podcast_episode',
+            'song',
+            'stream',
+            'tvshow',
+            'tvshow_season',
+            'user',
+            'video'
         );
 
         if ($object_type !== null) {